Make build Linux only for now
Sadly, nothing that I tried seems to work when it comes to running
xvfb ONLY for linux and not for osx. It just fails when a condition
is being put in, either in the travis.yml or by being invoked from
a simple bash script (see the .travis folder).

For now the tests will only be running on linux machines and maybe
someone would have a better idea of what is going wrong with xvfb
later on.